+++ apps/google-docs/src/locations/Page/components/review/mapping/ViewMappingCard.tsx
@@ -1,6 +1,6 @@
 import { Box } from '@contentful/f36-components';
 import tokens from '@contentful/f36-tokens';
-import { FIELD_TYPE_SEPARATOR, truncateFieldPart } from './mappingCardText';
+import { FIELD_TYPE_SEPARATOR, truncateFieldPart } from './mappingCardTextUtils';
 import { TruncatedRow } from './TruncatedRow';
 
 const CONTENT_TYPE_MAX_LENGTH = 30;
